Metaverse/ms-dataquality-p/requirements.txt
aiohttp==3.10.10
boto3==1.35.40
botocore==1.35.40
fastapi==0.115.2
Flask==3.0.3
Flask_Cors==5.0.0
google_api_python_client==2.149.0
pandas==2.2.3
pandasql==0.7.3
protobuf==5.28.2
pydantic==2.9.2
pyspark==3.5.3
python-dotenv==1.0.1
python_gitlab==4.8.0
Requests==2.32.3
streamlit==1.39.0
uvicorn==0.32.0
cryptography==44.0.0